adding cors support and cleaning code a bit
