+++ Cargo.toml
@@ -2,10 +2,11 @@
 edition = "2021"
 name = "imgopt"
 authors = ["mpw <mariano@holaplex.com>"]
-version = "0.2.0"
+version = "0.2.1"
 
 [dependencies]
 actix-web = "4.0.1"
+actix-cors = "0.6.4"
 awc = { version = "3.0.1", features = [ "rustls" ] }
 env_logger = "0.9.1"
 image = "0.24.4"
@@ -29,5 +30,3 @@ resize = "0.7.2"
 png = "0.16.7"
 rgb = "0.8.24"
 sha1_smol = "1.0.0"
-#actix-web-httpauth = "0.6.0"
-#actix-cors = "0.6.1"
