+++ config-sample.toml
@@ -16,7 +16,6 @@ user_agent = "imgopt/0.2.0"
 storage_path = "storage"
 #Limit which width scaling options are available (optional)
 allowed_sizes = [ 50, 100, 200, 400, 500, 600, 800, 700, 900, 1000, 1100, 1200, 1300, 1400, 1500 ]
-
 #max age header for twitter (in seconds) (optional)
 twitter.cache.max_age = 3600
 #Set to true will enable the query parameter ?url= to cache and process an asset found in the provided endpoint
